fix ambiguous steps about DB
